+++ src/imio/smartweb/core/upgrades.py
@@ -62,3 +62,8 @@ def update_icons_and_names(context):
 def reload_registry(context):
     portal_setup = api.portal.get_tool("portal_setup")
     portal_setup.runImportStepFromProfile(PROFILEID, "plone.app.registry")
+
+
+def reload_types(context):
+    portal_setup = api.portal.get_tool("portal_setup")
+    portal_setup.runImportStepFromProfile(PROFILEID, "typeinfo")