+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 1.0a18 (unreleased)
 -------------------
 
+- Fix footer viewlet markup to be included in Plone footer
+  [laulaz]
+
 - Add faceted layout class to body if a faceted layout is define.
   [boulch]